refactor: typing test
-Consolidate components like StartMenu and TestTimeOptions into SpeedTest component since it isn't being used anywhere else. -Add minor changes to TypingStats and GameOverMenu components to prepare for integration into Lesson component.
