--- client/src/components/forms/TestTimeOptions.tsx
@@ -1,39 +1,0 @@
-import styles from "./styles/StartMenu.module.css";
-import Min from "../svg/Min";
-
-interface PropTypes {
-  [key: string]: string[];
-}
-
-//Used by StartMenu.tsx component
-function TestTimeOptions({ timeOptions }: PropTypes) {
-  return (
-    <ul className="relative my-8 grid min-h-[10em] w-full max-w-[450px] grid-cols-3 gap-x-2 gap-y-8 text-xl sm:min-h-[4em] sm:w-11/12 sm:max-w-none sm:grid-cols-5 sm:justify-evenly sm:gap-y-0 sm:text-2xl">
-      {timeOptions.map((time: string, index: number) => (
-        <li key={index} className={`flex items-center justify-center `}>
-          <input
-            id={`radio-${time}`}
-            type="radio"
-            name="time-setting"
-            className="opacity-1 absolute"
-            defaultChecked={index === 0 ? true : false}
-            value={time}
-          />
-          <label
-            htmlFor={`radio-${time}`}
-            className={`${styles["menu-label"]} z-[1] flex h-[4.1em] w-[4.1em] flex-col items-center justify-center rounded-lg border-2 border-slate-200 bg-white fill-slate-500 text-slate-500 outline-default-sky-blue hover:cursor-pointer hover:border-sky-600 hover:fill-sky-700 hover:font-medium hover:text-sky-700 sm:h-24 sm:w-24`}
-          >
-            <span className="relative font-bold">{time}</span>
-            <span
-              className={`${styles["svg-text"]} relative text-xl sm:text-2xl`}
-            >
-              <Min />
-            </span>
-          </label>
-        </li>
-      ))}
-    </ul>
-  );
-}
-
-export default TestTimeOptions;
