+++ client/src/components/layout/BestStats.tsx
@@ -53,7 +53,7 @@ function BestStats({ userId, difficultyLevel, testName, gameOver }: PropType) {
     >
       {gameOver && (
         <>
-          <button className="absolute bottom-0 flex h-10 w-full items-center justify-center gap-3 bg-black  tracking-widest text-sky-300 opacity-80 transition-all group-hover:h-12 group-hover:text-sky-200 group-hover:opacity-90">
+          <button className="absolute bottom-0 flex h-10 w-full items-center justify-center gap-3 bg-defaultblue  tracking-widest text-sky-300 opacity-80 transition-all group-hover:h-12 group-hover:text-sky-200 group-hover:opacity-90">
             <span>
               <Icon
                 title="settings-icon"
